+++ raptorq/src/constraint_matrix.rs
@@ -1,8 +1,13 @@
+#[cfg(feature = "std")]
+use std::vec::Vec;
+
+#[cfg(not(feature = "std"))]
+use alloc::vec::Vec;
+
 use crate::base::intermediate_tuple;
 use crate::matrix::BinaryMatrix;
 use crate::octet::Octet;
 use crate::octet_matrix::DenseOctetMatrix;
-use crate::octets::{add_assign, fused_addassign_mul_scalar};
 use crate::rng::rand;
 use crate::systematic_constants::extended_source_block_symbols;
 use crate::systematic_constants::num_hdpc_symbols;
@@ -59,67 +64,48 @@ pub fn enc_indices(
 #[allow(non_snake_case)]
 fn generate_hdpc_rows(Kprime: usize, S: usize, H: usize) -> DenseOctetMatrix {
     let mut matrix = DenseOctetMatrix::new(H, Kprime + S + H, 0);
-    // G_HDPC
+    // Compute G_HDPC using recursive formulation, since this is much faster than a
+    // naive matrix multiplication approach
 
-    // Generates the MT matrix
-    // See section 5.3.3.3
-    let mut mt: Vec<Vec<u8>> = vec![vec![0; Kprime + S]; H];
+    let mut result: Vec<Vec<u8>> = vec![vec![0; Kprime + S]; H];
+    // Initialize the last column to alpha^i, which comes from multiplying the last column of MT
+    // with the lower right 1 in GAMMA
+    #[allow(clippy::needless_range_loop)]
+    for i in 0..H {
+        result[i][Kprime + S - 1] = Octet::alpha(i).byte();
+    }
+
+    // Now compute the rest of G_HDPC.
+    // Note that for each row in GAMMA, i'th col = alpha * (i + 1)'th col
+    // Therefore we can compute this right to left, by multiplying by alpha each time, and adding
+    // the Rand() entries which will be associatively handled
+    for j in (0..=(Kprime + S - 2)).rev() {
+        #[allow(clippy::needless_range_loop)]
+        for i in 0..H {
+            result[i][j] = (Octet::alpha(1) * Octet::new(result[i][j + 1])).byte();
+        }
+        let rand6 = rand((j + 1) as u32, 6u32, H as u32) as usize;
+        let rand7 = rand((j + 1) as u32, 7u32, (H - 1) as u32) as usize;
+        let i1 = rand6;
+        let i2 = (rand6 + rand7 + 1) % H;
+        result[i1][j] ^= Octet::one().byte();
+        result[i2][j] ^= Octet::one().byte();
+    }
+
+    // Copy G_HDPC into matrix
     #[allow(clippy::needless_range_loop)]
     for i in 0..H {
         #[allow(clippy::needless_range_loop)]
-        for j in 0..=(Kprime + S - 2) {
-            let rand6 = rand((j + 1) as u32, 6u32, H as u32) as usize;
-            let rand7 = rand((j + 1) as u32, 7u32, (H - 1) as u32) as usize;
-            if i == rand6 || i == (rand6 + rand7 + 1) % H {
-                mt[i][j] = 1;
-            }
-        }
-        mt[i][Kprime + S - 1] = Octet::alpha(i).byte();
-    }
-    // Multiply by the GAMMA matrix
-    // See section 5.3.3.3
-    let mut gamma_row = vec![0; Kprime + S];
-    // We only create the last row of the GAMMA matrix, as all preceding rows are just a shift left
-    #[allow(clippy::needless_range_loop)]
-    for j in 0..(Kprime + S) {
-        // The spec says "alpha ^^ (i-j)". However, this clearly can overflow since alpha() is
-        // only defined up to input < 256. Since alpha() only has 255 unique values, we must
-        // take the input mod 255. Without this the constraint matrix ends up being singular
-        // for 1698 and 8837 source symbols.
-        gamma_row[j] = Octet::alpha((Kprime + S - 1 - j) % 255).byte();
-    }
-    #[allow(clippy::needless_range_loop)]
-    for i in 0..H {
-        let mut result_row = vec![0; Kprime + S];
         for j in 0..(Kprime + S) {
-            let scalar = Octet::new(mt[i][j]);
-            if scalar == Octet::zero() {
-                continue;
-            }
-            if scalar == Octet::one() {
-                add_assign(
-                    &mut result_row[0..=j],
-                    &gamma_row[(Kprime + S - j - 1)..(Kprime + S)],
-                );
-            } else {
-                fused_addassign_mul_scalar(
-                    &mut result_row[0..=j],
-                    &gamma_row[(Kprime + S - j - 1)..(Kprime + S)],
-                    &scalar,
-                );
-            }
-        }
-        #[allow(clippy::needless_range_loop)]
-        for j in 0..(Kprime + S) {
-            if result_row[j] != 0 {
-                matrix.set(i, j, Octet::new(result_row[j]));
+            if result[i][j] != 0 {
+                matrix.set(i, j, Octet::new(result[i][j]));
             }
         }
     }
 
     // I_H
     for i in 0..H {
-        matrix.set(i, i + (Kprime + S) as usize, Octet::one());
+        matrix.set(i, i + (Kprime + S), Octet::one());
     }
 
     matrix
@@ -133,7 +119,6 @@ pub fn generate_constraint_matrix<T: BinaryMatrix>(
     source_block_symbols: u32,
     encoded_symbol_indices: &[u32],
 ) -> (T, DenseOctetMatrix) {
-
     let Kprime = extended_source_block_symbols(source_block_symbols) as usize;
     let S = num_ldpc_symbols(source_block_symbols) as usize;
     let H = num_hdpc_symbols(source_block_symbols) as usize;
@@ -162,7 +147,7 @@ pub fn generate_constraint_matrix<T: BinaryMatrix>(
 
     // I_S
     for i in 0..S {
-        matrix.set(i as usize, i + B as usize, Octet::one());
+        matrix.set(i, i + B, Octet::one());
     }
 
     // G_LDPC,2
@@ -182,9 +167,10 @@ pub fn generate_constraint_matrix<T: BinaryMatrix>(
         let tuple = intermediate_tuple(i, lt_symbols, sys_index, p1);
 
         for j in enc_indices(tuple, lt_symbols, pi_symbols, p1) {
-            matrix.set(row as usize + S + H, j, Octet::one());
+            matrix.set(row + S + H, j, Octet::one());
         }
     }
 
     (matrix, generate_hdpc_rows(Kprime, S, H))
 }
+