+++ raptorq/src/symbol.rs
@@ -1,12 +1,22 @@
+#[cfg(feature = "std")]
+use std::{ops::AddAssign, vec::Vec};
+
+#[cfg(not(feature = "std"))]
+use core::ops::AddAssign;
+
+#[cfg(not(feature = "std"))]
+use alloc::vec::Vec;
+
 use crate::octet::Octet;
 use crate::octets::add_assign;
 use crate::octets::fused_addassign_mul_scalar;
 use crate::octets::mulassign_scalar;
+#[cfg(feature = "serde_support")]
 use serde::{Deserialize, Serialize};
-use std::ops::AddAssign;
 
 /// Elementary unit of data, for encoding/decoding purposes.
-#[derive(Clone, Debug, PartialEq, PartialOrd, Eq, Ord, Serialize, Deserialize, Hash)]
+#[derive(Clone, Debug, PartialEq, PartialOrd, Eq, Ord, Hash)]
+#[cfg_attr(feature = "serde_support", derive(Serialize, Deserialize))]
 pub struct Symbol {
     value: Vec<u8>,
 }